correction

+++ 0x02-functions_nested_loops/103-fibonacci.c
@@ -8,7 +8,7 @@ int main(void)
 {
 	long a;
 	long b;
-	int sum;
+	long sum;
 
 	a = 1;
 	b = 2;
